v0.5.0

+++ intg-requests/config.py
@@ -18,6 +18,7 @@ class Setup:
         "setup_reconfigure": False,
         "bundle_mode": False,
         "cfg_path": "config.json",
+        "tcp_text_timeout": 2,
         "rq_timeout": 2,
         "rq_ssl_verify": True,
         "rq_user_agent": "uc-intg-requests",
@@ -35,12 +36,14 @@ class Setup:
         "id-head": "http-head",
         "name-head": "HTTP Head",
         "id-wol": "wol",
-        "name-wol": "Wake on LAN"
+        "name-wol": "Wake on LAN",
+        "id-tcp-text": "tcp-text",
+        "name-tcp-text": "Text over TCP"
     }
-    __setters = ["standby", "setup_complete", "setup_reconfigure", "rq_timeout", "rq_ssl_verify", "rq_fire_and_forget", "bundle_mode", "cfg_path"]
-    __storers = ["setup_complete", "rq_timeout", "rq_ssl_verify", "rq_fire_and_forget"] #Skip runtime only related values in config file
+    __setters = ["standby", "setup_complete", "setup_reconfigure", "tcp_text_timeout", "rq_timeout", "rq_user_agent", "rq_ssl_verify", "rq_fire_and_forget", "bundle_mode", "cfg_path"]
+    __storers = ["setup_complete", "tcp_text_timeout", "rq_timeout", "rq_user_agent", "rq_ssl_verify", "rq_fire_and_forget"] #Skip runtime only related values in config file
 
-    all_cmds = ["get", "post", "patch", "put", "delete", "head", "wol"]
+    all_cmds = ["get", "post", "patch", "put", "delete", "head", "wol", "tcp-text"]
     rq_ids = [__conf["id-get"], __conf["id-post"], __conf["id-patch"], __conf["id-put"], __conf["id-delete"], __conf["id-head"]]
     rq_names = [__conf["name-get"], __conf["name-post"], __conf["name-patch"], __conf["name-put"], __conf["name-delete"], __conf["name-head"]]
 
@@ -54,46 +57,51 @@ class Setup:
 
     @staticmethod
     def set(key, value):
-        """Set and store a value for the specified key into the runtime storage and config file.
-        Storing setup_complete flag during reconfiguration will be ignored"""
+        """Save a value for the specified key into the runtime storage and store some of them into the config file.
+        Saving and storing setup_complete flag during reconfiguration will be ignored as well as storing values that have not been changed from the default"""
         if key in Setup.__setters:
             if Setup.__conf["setup_reconfigure"] is True and key == "setup_complete":
-                _LOG.debug("Ignore setting and storing setup_complete flag during reconfiguration")
+                _LOG.debug("Ignore saving and storing setup_complete flag during reconfiguration")
             else:
+                skip_store = False
+                if key in Setup.__storers and value == Setup.__conf[key]:
+                    skip_store = True
+                    _LOG.debug("Skip storing " + key + ": " + str(value) + " into config file as it has not been changed from the default value of " + str(Setup.__conf[key]))
+
                 Setup.__conf[key] = value
-                _LOG.debug("Stored " + key + ": " + str(value) + " into runtime storage")
+                _LOG.debug("Saved " + key + ": " + str(value) + " into runtime storage")
 
                 #Store key/value pair in config file
-                if key in Setup.__storers:
+                if skip_store is False:
+                    if key in Setup.__storers:
+                        jsondata = {key: value}
+                        if os.path.isfile(Setup.__conf["cfg_path"]):
+                            try:
+                                with open(Setup.__conf["cfg_path"], "r+", encoding="utf-8") as f:
+                                    l = json.load(f)
+                                    l.update(jsondata)
+                                    f.seek(0)
+                                    f.truncate() #Needed when the new value has less characters than the old value (e.g. false to true or 11 to 5 seconds timeout)
+                                    json.dump(l, f)
+                                    _LOG.debug("Stored " + key + ": " + str(value) + " into " + Setup.__conf["cfg_path"])
+                            except OSError as o:
+                                raise Exception(o) from o
+                            except Exception as e:
+                                raise Exception("Error while storing " + key + ": " + str(value) + " into " + Setup.__conf["cfg_path"]) from e
 
-                    jsondata = {key: value}
-                    if os.path.isfile(Setup.__conf["cfg_path"]):
-                        try:
-                            with open(Setup.__conf["cfg_path"], "r+", encoding="utf-8") as f:
-                                l = json.load(f)
-                                l.update(jsondata)
-                                f.seek(0)
-                                f.truncate() #Needed when the new value has less characters than the old value (e.g. false to true or 11 to 5 seconds timeout)
-                                json.dump(l, f)
-                                _LOG.debug("Stored " + key + ": " + str(value) + " into " + Setup.__conf["cfg_path"])
-                        except OSError as o:
-                            raise Exception(o) from o
-                        except Exception as e:
-                            raise Exception("Error while storing " + key + ": " + str(value) + " into " + Setup.__conf["cfg_path"]) from e
+                        #Create config file first if it doesn't exists yet
+                        else:
+                            try:
+                                with open(Setup.__conf["cfg_path"], "w", encoding="utf-8") as f:
+                                    json.dump(jsondata, f)
+                                _LOG.debug("Created " + Setup.__conf["cfg_path"] + " and stored " + key + ": " + str(value) + " in it")
+                            except OSError as o:
+                                raise Exception(o) from o
+                            except Exception as e:
+                                raise Exception("Error while storing " + key + ": " + str(value) + " into " + Setup.__conf["cfg_path"]) from e
 
-                    #Create config file first if it doesn't exists yet
                     else:
-                        try:
-                            with open(Setup.__conf["cfg_path"], "w", encoding="utf-8") as f:
-                                json.dump(jsondata, f)
-                            _LOG.debug("Created " + Setup.__conf["cfg_path"] + " and stored " + key + ": " + str(value) + " in it")
-                        except OSError as o:
-                            raise Exception(o) from o
-                        except Exception as e:
-                            raise Exception("Error while storing " + key + ": " + str(value) + " into " + Setup.__conf["cfg_path"]) from e
-
-                else:
-                    _LOG.debug(key + " not found in __storers because it should not be stored in the config file")
+                        _LOG.debug(key + " not found in __storers because it should not be stored in the config file")
 
         else:
             raise NameError(key + " not found in __setters because it should not be changed")
@@ -117,26 +125,40 @@ class Setup:
             if not Setup.__conf["setup_complete"]:
                 _LOG.warning("The setup was not completed the last time. Please restart the setup process")
             else:
+                if "tcp_text_timeout" in configfile:
+                    Setup.__conf["tcp_text_timeout"] = configfile["tcp_text_timeout"]
+                    _LOG.info("Loaded custom text over tcp timeout of " + str(configfile["tcp_text_timeout"]) + " seconds into runtime storage from " + Setup.__conf["cfg_path"])
+                else:
+                    _LOG.info("Skip loading custom text over tcp timeout as it has not been changed during setup. \
+The Default value of " + str(Setup.get("tcp_text_timeout")) + " seconds will be used")
+                
+                if "rq_user_agent" in configfile:
+                    Setup.__conf["rq_user_agent"] = configfile["rq_user_agent"]
+                    _LOG.info("Loaded custom http requests user agent " + str(configfile["rq_user_agent"]) + " into runtime storage from " + Setup.__conf["cfg_path"])
+                else:
+                    _LOG.info("Skip loading custom http requests user agent as it has not been changed during setup. \
+The Default user agent " + str(Setup.get("rq_user_agent")) + " will be used")
+
                 if "rq_timeout" in configfile:
                     Setup.__conf["rq_timeout"] = configfile["rq_timeout"]
-                    _LOG.info("Loaded requests timeout of " + str(configfile["rq_timeout"]) + " seconds into runtime storage from " + Setup.__conf["cfg_path"])
+                    _LOG.info("Loaded custom http requests timeout of " + str(configfile["rq_timeout"]) + " seconds into runtime storage from " + Setup.__conf["cfg_path"])
                 else:
-                    _LOG.info("Skip loading custom requests timeout as it has not been changed during setup. \
-                    Default value of " + str(Setup.get("rq_timeout")) + " seconds will be used")
+                    _LOG.info("Skip loading custom http requests timeout as it has not been changed during setup. \
+The Default value of " + str(Setup.get("rq_timeout")) + " seconds will be used")
 
                 if "rq_ssl_verify" in configfile:
                     Setup.__conf["rq_ssl_verify"] = configfile["rq_ssl_verify"]
-                    _LOG.info("Loaded http ssl verification: " + str(configfile["rq_ssl_verify"]) + " into runtime storage from " + Setup.__conf["cfg_path"])
+                    _LOG.info("Loaded custom http ssl verification: " + str(configfile["rq_ssl_verify"]) + " into runtime storage from " + Setup.__conf["cfg_path"])
                 else:
                     _LOG.info("Skip loading http ssl verification flag as it has not been changed during setup. \
-                    Default value " + str(Setup.get("rq_ssl_verify")) + " will be used")
+The Default value " + str(Setup.get("rq_ssl_verify")) + " will be used")
 
                 if "rq_fire_and_forget" in configfile:
                     Setup.__conf["rq_fire_and_forget"] = configfile["rq_fire_and_forget"]
-                    _LOG.info("Loaded fire_and_forget: " + str(configfile["rq_fire_and_forget"]) + " into runtime storage from " + Setup.__conf["cfg_path"])
+                    _LOG.info("Loaded custom fire_and_forget: " + str(configfile["rq_fire_and_forget"]) + " flag into runtime storage from " + Setup.__conf["cfg_path"])
                 else:
                     _LOG.info("Skip loading fire_and_forget as it has not been changed during setup. \
-                    Default value " + str(Setup.get("rq_fire_and_forget")) + " will be used")
+The Default value " + str(Setup.get("rq_fire_and_forget")) + " will be used")
 
         else:
             _LOG.info(Setup.__conf["cfg_path"] + " does not exist (yet). Please start the setup process")
